move npmrc io
