+++ src/services/put-npm-auth-token.ts
@@ -1,6 +1,6 @@
 import { emptyNpmrc, getHomeNpmrcPath, setToken } from "../domain/npmrc";
 import { RegistryUrl } from "../domain/registry-url";
-import { saveNpmrcUsing, tryLoadNpmrcUsing } from "../io/npmrc-io";
+import { saveNpmrcUsing, tryLoadNpmrcUsing } from "./npmrc-io";
 import {
   readTextFile,
   writeTextFile,
